GUI: add group tables

+++ spinfood/src/main/java/model/CSVFileReader.java
@@ -147,7 +147,7 @@ public class CSVFileReader extends ParticipantManager implements FileReader {
      * @param kitchen
      */
     private void createAndAddParticipant(String id, String name, double age, HasKitchen hasKitchen, FoodPreference foodPreference, Sex sex, Kitchen kitchen) {
-        participants.add(new Participant(id, name, age, hasKitchen, foodPreference, sex, kitchen));
+        participants.add(new Participant(id, name, (int)age, hasKitchen, foodPreference, sex, kitchen));
     }
 
     /**
@@ -166,8 +166,8 @@ public class CSVFileReader extends ParticipantManager implements FileReader {
      */
     private void createAndAddCSVPair(String id1, String name1, double age1, HasKitchen hasKitchen1, FoodPreference foodPreference1,
                                      Sex sex1, Kitchen kitchen, String id2, String name2, double age2, Sex sex2) {
-        Participant person1 = new Participant(id1, name1, age1, hasKitchen1, foodPreference1, sex1, kitchen);
-        Participant person2 = new Participant(id2, name2, age2, hasKitchen1, foodPreference1, sex2, kitchen);
+        Participant person1 = new Participant(id1, name1, (int)age1, hasKitchen1, foodPreference1, sex1, kitchen);
+        Participant person2 = new Participant(id2, name2,(int) age2, hasKitchen1, foodPreference1, sex2, kitchen);
         CSV_Pairs.add(new Pair(person1, person2, true));
     }
